backend/pyproject.toml
[project]
name = "fastapi-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "email-validator>=2.3.0",
    "fastapi-mail>=1.6.2",
    "fastapi[standard]>=0.129.0",
    "hashids>=1.3.1",
    "passlib>=1.7.4",
    "python-jose[cryptography]>=3.5.0",
    "sqlmodel>=0.0.34",
    "taskipy>=1.14.1",
    "slowapi>=0.1.9",
    "pyotp>=2.9.0",
    "qrcode>=8.0",
    "pillow>=11.0.0",
    "aiosqlite>=0.20.0",
    "asyncpg>=0.30.0",
    "greenlet>=3.3.1",
    "alembic>=1.18.4",
    "bcrypt>=4.0.0,<5.0.0",
    "casbin>=1.37.0",
    "casbin-sqlalchemy-adapter>=0.6.0",
    "casbin-async-sqlalchemy-adapter>=0.3.3",
    "celery>=5.4.0",
    "redis>=5.2.0",
]

[dependency-groups]
dev = [
    "watchfiles>=1.1.1",
]
test = [
    "factory-boy>=3.3.3",
    "httpx>=0.28.1",
    "hypothesis>=6.151.9",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
]

[tool.taskipy.tasks]
start = "uvicorn src.main:app --reload"
test = "pytest --cov=app tests/"
initial-migrate = 'alembic revision --autogenerate -m "init"'
migrate = "alembic upgrade head"
celery = "celery -A src.apps.core.celery_app worker --loglevel=info"
